Primera version estable (#10)
* Eliminacion de archivos locales del proyecto
Se agregan algunos paquetes nuevos: interfaces, frames, db
Se agregar una base de datos temporal en memoria
Actualizacion de estructura y forma en algunas partes

* Eliminacion de archivos de compilacion locales

* Actualizacion de gitignore

* Actualizacion nombre proyecto

* Eliminacion de comentarios

* Agregada funcionalidad de registro de cliente

* Mensaje de confirmacion

* Actualizacion de version a 1.1.0-beta

* agrego la conexion a base de datos

connect to DB class

* Se agrega el registro de reservacion (#5)

Se agregan las interfaces para los servicios de reservacion y cliente
Se elimina codigo muerto
Se mejora el flujo de uso


* updated the author

updated the author

* Actualizacion de version

* Integracion con la base de datos real (#7)

* Integracion con la base de datos real

* Eliminacion de comentarios

---------


* Se agrega inicio de sesion (#8)


* Se agrega la pantalla de reservacion existente (#9)

* Se agrega la pantall de reservacion existente

* Version bump

---------


* Actualizacion de version

---------

+++ src/rent/a/car/cliente/servidor/modelos/Reservacion.java
@@ -1,22 +1,14 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
 package rent.a.car.cliente.servidor.modelos;
 
-/**
- *
- * @author daniel.guzman
- */
 public class Reservacion {
 
     private Integer id;
     private int dias;
-    private int costo;
+    private double costo;
     private int vehiculo_fk;
     private int cliente_fk;
 
-    public Reservacion(Integer id, int dias, int costo, int vehiculo_fk, int cliente_fk) {
+    public Reservacion(Integer id, int dias, double costo, int vehiculo_fk, int cliente_fk) {
         this.id = id;
         this.dias = dias;
         this.costo = costo;
@@ -40,11 +32,11 @@ public class Reservacion {
         this.dias = dias;
     }
 
-    public int getCosto() {
+    public double getCosto() {
         return costo;
     }
 
-    public void setCosto(int costo) {
+    public void setCosto(double costo) {
         this.costo = costo;
     }
 
@@ -64,5 +56,4 @@ public class Reservacion {
         this.cliente_fk = cliente_fk;
     }
 
-    
 }
